+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.dao.MenuGroupRepository;
+import kitchenpos.domain.menugroup.MenuGroupRepository;
 import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.ui.request.MenuGroupCreateRequest;
 import kitchenpos.ui.response.MenuGroupResponse;
