+++ src/main/java/kitchenpos/application/ProductService.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.dao.ProductRepository;
+import kitchenpos.domain.product.ProductRepository;
 import kitchenpos.domain.product.Product;
 import kitchenpos.ui.request.ProductCreateRequest;
 import kitchenpos.ui.response.ProductResponse;
