+++ src/main/java/kitchenpos/application/TableService.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
-import kitchenpos.dao.OrderRepository;
-import kitchenpos.dao.OrderTableRepository;
+import kitchenpos.domain.order.OrderRepository;
+import kitchenpos.domain.ordertable.OrderTableRepository;
 import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.ui.request.TableCreateRequest;
 import kitchenpos.ui.request.TableUpdateEmptyRequest;
