+++ src/main/java/kitchenpos/domain/menu/MenuProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.domain.menu;
 
 import kitchenpos.domain.menu.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
