+++ src/main/java/kitchenpos/domain/menu/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.domain.menu;
 
 import kitchenpos.domain.menu.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
