+++ src/main/java/kitchenpos/domain/order/OrderLineItemRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.domain.order;
 
 import kitchenpos.domain.order.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
