+++ src/main/java/kitchenpos/domain/ordertable/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.domain.ordertable;
 
 import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
