+++ src/main/java/kitchenpos/domain/product/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.domain.product;
 
 import kitchenpos.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
